+++ packages/components/src/styling/src/styled-system/index.ts
@@ -1,1 +1,4 @@
 export * from "./useStyledSystem";
+export * from "./StylingContext";
+export * from "./props-handler";
+export * from "./theme-vars-utils";
