+++ packages/components/src/styling/src/theming/index.ts
@@ -1,3 +1,2 @@
-export * from "./createThemeVars";
 export * from "./ThemeProvider";
 export * from "./useThemeComputedStyle";
